backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prospector"
version = "1.0.0"
description = "Sistema de prospecção B2B automatizado — Composto Web"
requires-python = ">=3.12"

dependencies = [
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",

    # Validação e configuração
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",

    # Banco de dados
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",

    # Celery + Redis
    "celery[redis]>=5.4.0",
    "redis[hiredis]>=5.2.0",

    # HTTP client async
    "httpx>=0.27.0",

    # OpenAI
    "openai>=1.55.0",

    # Google Gemini SDK (GA desde maio/2025)
    # NÃO usar google-generativeai — descontinuado em novembro/2025
    "google-genai>=1.0.0",

    # Speechify / provedor de voz
    # ElevenLabs:  pip install elevenlabs>=1.9.0
    # PlayHT:      pip install pyht>=0.1.0
    # OpenAI TTS já incluso no openai SDK acima

    # Autenticação
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # Utilitários
    "python-dotenv>=1.0.1",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.1",
    "faker>=30.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pyright]
pythonVersion = "3.12"
# pytest-asyncio não fornece stubs completos; ignora decoradores sem tipo
reportUntypedFunctionDecorator = false
# structlog usa magic methods que não possuem stubs; pacote instalado em runtime
reportMissingModuleSource = false
reportMissingImports = false
# Celery .delay() não está tipado nos stubs do celery
reportAttributeAccessIssue = false
# structlog — tipos dinâmicos
reportUnknownMemberType = false
reportUnknownVariableType = false
